pyproject.toml
[project]
name = "ner-extraction"
version = "0.1.0"
description = "NER extraction using Prompt Engineering, RAG, Fine-tuning, and LangExtract"
readme = "PROJECT_README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.35.0",
    "datasets>=2.14.0",
    "peft>=0.5.0",
    "bitsandbytes>=0.41.0",
    "accelerate>=0.24.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    "faiss-cpu>=1.7.4",
    "sentence-transformers>=2.2.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "tqdm>=4.65.0",
    "scikit-learn>=1.3.0",
    "jupyter>=1.0.0",
    "ipywidgets>=8.0.0",
    "tabulate>=0.9.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "langextract>=0.1.0",
    "google-generativeai>=0.3.0",
    "newspaper3k>=0.2.8",
    "google-genai>=1.52.0",
    "loguru>=0.7.3",
    "vllm>=0.11.2",
    "outlines>=1.2.8",
]
